Agregue la clase abstracta Usuario y sus herencias además de la clase penalidad
