+++ src/Modelo/Recepcionista.java
@@ -1,0 +1,24 @@
+package Modelo;
+
+/**
+ *
+ * @author Dayanna
+ */
+public class Recepcionista extends Usuario {
+    public Recepcionista(String id_codigo, String contraseña, String nombre, String apellido) {
+        this.id_codigo = id_codigo;
+        this.contraseña = contraseña;
+        this.nombre = nombre;
+        this.apellido = apellido;
+        this.tipoDeUser = "Recepcionista";
+    }
+
+    public void registrarDevolucion() {
+        System.out.println("Devolución registrada correctamente.");
+    }
+
+    public void mostrarInfo() {
+        System.out.println("Recepcionista: " + nombre + " " + apellido);
+    }
+}
+
